LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_setpointercapture_inactive_button_mouse-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
setPointerCapture

Test Description: This test checks if setPointerCapture works properly.
Put your mouse over the black rectangle
Move you mouse out to complete the test

Pointer Events setPointerCapture Tests

The following pointer types were detected: (none).


FAIL setPointerCapture + inactive button state Unhandled rejection: unimplemented
